Q&A-LLM-with-LangchainProject-3/requirements.txt
langchain 
openai
huggingface_hub
python-dotenv
streamlit